Improve usability of unmanaged namespaces
This change makes it a nondestructive operation to switch a namespace
from managed to unmanaged and back, by _not_ deleting the HC or anchors
while a namespace is unmanaged. It also fixes one serious bug
(namespaces could not be changed to roots because the unnamed namespace
didn't match the regex) and adds a lot more details in the error
messages. Finally, it allows users to fix broken hierarchies by removing
a formerly-managed parent.

I also changed the terminology from "excluded" to "unmanaged"
throughout.

Testing: This is hard to test automatically since it requires constant
redeployments of HNC. So I did extensive manual testing instead:

* Try to arrange hierarchies with unmanaged namespace, both excluded
(e.g hnc-system) and ones that don't match the regex. Verify that the
error messages clearly indicate the correct problem.
* Create a legal hierarchy and then break it by changing the regex.
Verify that I can fix it without restarting HNC.
* Create a legal hierarchy, break it with the regex, and then restore it
with the regex. Verify that the original hierarchy comes back.
* Verify that anchors and hierarchyconfigs lose their finalizers when
their namespaces become unmanaged, and get them back if they become
managed again.