Initial implementation of managed labels
See #47. This is the initial implementation of managed labels and
annotations - that is, the ability to set a label (or annotation) in a
HierarchyConfiguration object, and have that label (...) propagated to
all descendants, similar to the way objects are propagated. As with
objects, only allowlisted labels are propagated, as defined by the
command line option '--managed-namespace-[label|annotation]'.

Still to come: validator support, better conditions for conflicts,
better testing for external namespaces, better testing for more regexes,
documentation, end-to-end tests.

Tested: see new integ tests.