Add cli argument `--disable-enforced-types`
This kinda implements the request made in #104, as it makes it possible
to disable the default enforced types `roles` and `rolebindings` with an
additional cli argument that can be used with the hnc-manager. The old
behaviour (using enforcedTypes) is the default.

Tested: added new e2e tests and checked them against a vanilla k8s
cluster. No issues found.