Fix timeout stopping testEnv api-server on K8s 1.20+
The source problem seems to be a pending watch request, initiated by the controller after the kube-apiserver has received SIGTERM.
The request does not appear to have a timeout, and is not answered by the apiserver.
Graceful shutdown introduced in 1.21 appears to wait for the pending request to terminate before shutting the apiserver completely.

To avoid the error, one must thus avoid sending watch requests after SIGTERMing the apiserver, or avoid the watch request pending for more than StopTimeout (default 20s).

Tested: Tried to run 'make test' before the fix (using envTest K8s 1.23.1), but that fails with the error "timeout waiting for process kube-apiserver to stop". After the fix all tests run successfully.