Add per-object-level filtering based on labels
The set of rules is hardcoded as of now, as proposed in #104. The
current rules are Rancher-specific, as Rancher does some propagation of
roles, rolebindings and serviceaccounts on its own via its "projects"
(used to group namespaces). Label equality is checked against both key
and value - no regex support for now.

Tested: added a new test to check whether roles / configmap propagation
is prevented if the label is matched.