Use DynamicRESTMapper to lookup metadata from apiserver
This is mainly a refactoring of existing code, but modifies the internal machinery slightly. The initial motivation was to add more apiserver lookups to improve webhook validations in the `objects` package. Since the existing code performing similar stuff was private to the `hncconfig` package I had to do some changes anyway.

It turns out that controller-runtime provides a `DynamicRESTMapper`, that I think will fit very well in HNC for looking up resource information from the apiserver. HNC needs dynamic information in runtime, and that is exactly what this class provides. It also has some additional configuration features that can eventually be used in a follow-up PR.

Testing: Ran all tests locally with success, but I do not currently have a cluster available to test this change more. I expect it to work well, but doing changes to the machinery is always a bit risky.