EXPERIMENTAL HA support
This change lets the HNC manager be run both without webhooks and
in webhook-only mode. This allows us to create two Deployments - one for
the controller which actually makes changes, and another replicated
deployment for the webhooks, each of which independently sync their
in-memory forests and serve webhook requests.

Taking advantage of this new feature requires considerable changes to
the manifests, which I'm checking in as a separate commit.

Tested: change passes smoke test when deployed with HA manifests, but I
haven't done any careful validation that (for example) the HA deployment
isn't actually writing anything.