Validate subnamespace anchors on update
Now that labels and annotations can be added to subnamespace anchors, these should now also be validated when they are updated.

Tested: before this change on K8s 1.16 I was able to modify the labels on subnamespace anchors without any validation running. With this change my modification is now validated and blocked.