add `staticcheck` to the make build process, #146
Add `staticcheck` to the build process, including a `check-tools` make task that installs `staticcheck` if necessary. `staticcheck` is a golang linter / static analysis tool that raises warnings and errors if it finds any problems in the codebase. The `staticcheck` step is also run during various `make` tasks (e.g. `build`) and *blocks the build if any errors or warnings* are encountered.

Tested: `make test` was run successfully.