Fix resources not managed by HNC being deleted
This commit fixes the problem that when HNC restarts the source objects
are deleted and then re-created.

When a source object is deleted the code now checks the decedents for
existing source objects and adds an annotation to them to make sure that
only sources hnc is aware of are deleted.

It also makes sure that all source files are synchronised with the
forest even if they are just propergated objects so that the forest is
more accurate reflection of what's on the cluster.