Refactor some HRQ code for readability
I found some of this code very difficult to read, and there was a bunch
of stuff in the 'utils' package that was being exported but not used
(this is because the library was originally copied from K8s, but it's
been heavily modified since then). Hopefully others find these cleanups
useful to.

I did simplify the TryUseResources function to allow decreasing usage.
This opens up a *very* slight race condition at the cost of making the
function more robust to future changes in K8s and significantly easier
to understand and test.