Add periodic HRQ resync
If the HRQ logic is correct, the incremental updates should always work,
but there's always the chance of a bug, so add a periodic (1m by
default) recalculation from scratch to ensure that the incremental
updates are working correctly. If a discrepancy is found, it dumps an
error in the log and updates the HRQs to the correct values.

Tested: broke the checker in various ways and ensured that the integ
test failed.