Add API to list resources under namespace tree
This change adds a new command and configuration to register an API
endpoint which will allow users to query for resources belonging to
descendants of any namespace. This means that finding all resources
under a parent namespace can be done in one request, instead finding and
iterating over all descendants. The new endpoints are:

* /apis/resources.hnc.x-k8s.io/v1alpha2
 - discovery endpoint
* /apis/resources.hnc.x-k8s.io/v1alpha2/{resource}
 - global resource query, equivalent to the same global request for the
   original resource
* /apis/resources.hnc.x-k8s.io/v1alpha2/namespaces/{namespace}/{resource}
 - returns resources in all namespaces under {namespace}, including
   {namespace}.