Update to Go 1.22
Bump the Go versions in various
files and run go mod tidy to clean
up things, Then update to
controller-tools 0.15 to fix
the segmentation fault when running
controller-gen.

Being able to run controller-gen
gives a number of new compilation
errors, so update the K8S packages
to 0.27.16, controller-runtime to
0.15.3 which supports Kubernetes 1.27,
and open-policy-agent/cert-controller
to 0.8.0 which also supports
Kubernetes 1.27.